feat(identity-registration): codereview
