+++ package.json
@@ -22,7 +22,7 @@
     "@react-hook/resize-observer": "^2.0.2",
     "@scure/base": "^1.2.4",
     "@scure/bip39": "^1.6.0",
-    "dash-core-sdk": "dash-core-sdk@1.1.3-dev.1",
+    "dash-core-sdk": "1.1.3-dev.1",
     "dash-platform-sdk": "1.3.2-dev.3",
     "dash-ui-kit": "2.1.0-dev",
     "eciesjs": "^0.4.15",
